fixed interfaces folder name and moved add post component to a separate file
