+++ client/src/Interfaces/Animal.ts
