+++ client/src/Validation/addPost.ts
@@ -1,0 +1,9 @@
+import * as yup from 'yup';
+
+const postSchema = yup.object().shape({
+  content: yup.string().min(120).required(),
+  image: yup.mixed(),
+  TagId: yup.number().required(),
+  AnimalId: yup.number().required(),
+});
+export default postSchema;
